correcion

+++ held-karp-serial/main.cpp
@@ -23,7 +23,7 @@ int main(){
 
 void getanswer(vector <vector<double> > grafo){
     int N = grafo.size();
-    vector< vector<double> > best( 1<<(N-1), vector<double>( N, INT8_MAX ) );
+    vector< vector<double> > best( 1<<(N-1), vector<double>( N, INT_MAX ) );
     for (int visited = 1; visited < (1<<(N-1)); ++visited) {
         for (int last = 0; last < (N-1); ++last) {
  
@@ -48,7 +48,7 @@ void getanswer(vector <vector<double> > grafo){
             }
         }
     }
-    double answer = INT8_MAX;
+    double answer = INT_MAX;
     for (int last=0; last<N-1; ++last) {
         answer = min( 
             answer,
